FIXED.with cart.

+++ books/src/main/java/soft/co/books/domain/service/AuthorService.java
@@ -8,12 +8,17 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.mongodb.core.MongoTemplate;
 import org.springframework.data.mongodb.core.query.Query;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import soft.co.books.configuration.Constants;
 import soft.co.books.configuration.database.CustomBaseService;
+import soft.co.books.configuration.error.CustomizeException;
 import soft.co.books.domain.collection.Author;
+import soft.co.books.domain.collection.Book;
 import soft.co.books.domain.repository.AuthorRepository;
 import soft.co.books.domain.service.dto.AuthorDTO;
 import soft.co.books.domain.service.dto.PageResultDTO;
 
+import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
 
@@ -24,17 +29,23 @@ import static org.springframework.data.mongodb.core.query.Criteria.where;
  * Service class for managing authors.
  */
 @Service
+@Transactional
 public class AuthorService extends CustomBaseService<Author, String> {
 
     private final Logger log = LoggerFactory.getLogger(AuthorService.class);
 
     private final AuthorRepository authorRepository;
 
+    private final CountryService countryService;
+
     private MongoTemplate mongoTemplate;
 
-    public AuthorService(AuthorRepository authorRepository, MongoTemplate mongoTemplate) {
+    public AuthorService(AuthorRepository authorRepository,
+                         CountryService countryService,
+                         MongoTemplate mongoTemplate) {
         super(authorRepository);
         this.authorRepository = authorRepository;
+        this.countryService = countryService;
         this.mongoTemplate = mongoTemplate;
     }
 
@@ -47,11 +58,11 @@ public class AuthorService extends CustomBaseService<Author, String> {
         query.with(pageable.getSort());
 
         if (authorDTO.getFirstName() != null && !authorDTO.getFirstName().isEmpty())
-            query.addCriteria(where("firstName").regex(authorDTO.getFirstName()));
+            query.addCriteria(where("firstName").regex(authorDTO.getFirstName(), "i"));
         if (authorDTO.getLastName() != null && !authorDTO.getLastName().isEmpty())
-            query.addCriteria(where("lastName").regex(authorDTO.getLastName()));
+            query.addCriteria(where("lastName").regex(authorDTO.getLastName(), "i"));
         if (authorDTO.getCity() != null && !authorDTO.getCity().isEmpty())
-            query.addCriteria(where("city").regex(authorDTO.getCity()));
+            query.addCriteria(where("city").regex(authorDTO.getCity(), "i"));
         if (authorDTO.getBornDate() != null && !authorDTO.getBornDate().isEmpty())
             query.addCriteria(where("bornDate").is(authorDTO.getBornDate()));
         if (authorDTO.getDeathDate() != null && !authorDTO.getDeathDate().isEmpty())
@@ -67,8 +78,11 @@ public class AuthorService extends CustomBaseService<Author, String> {
         Author author = new Author();
         author.setFirstName(authorDTO.getFirstName());
         author.setLastName(authorDTO.getLastName());
+        author.setFullName(authorDTO.getFirstName().toLowerCase() + " " + authorDTO.getLastName().toLowerCase());
         author.setId(authorDTO.getId());
         author.setCity(authorDTO.getCity());
+        author.setState(authorDTO.getState());
+        author.setCountry(countryService.findOne(authorDTO.getCountry().getId()).get());
         author.setBornDate(authorDTO.getBornDate());
         author.setDeathDate(authorDTO.getDeathDate());
 
@@ -91,7 +105,10 @@ public class AuthorService extends CustomBaseService<Author, String> {
                 .map(author -> {
                     author.setFirstName(authorDTO.getFirstName());
                     author.setLastName(authorDTO.getLastName());
+                    author.setFullName(authorDTO.getFirstName().toLowerCase() + " " + authorDTO.getLastName().toLowerCase());
                     author.setCity(authorDTO.getCity());
+                    author.setState(authorDTO.getState());
+                    author.setCountry(countryService.findOne(authorDTO.getCountry().getId()).get());
                     author.setDeathDate(authorDTO.getDeathDate());
                     author.setBornDate(authorDTO.getBornDate());
                     authorRepository.save(author);
@@ -102,9 +119,19 @@ public class AuthorService extends CustomBaseService<Author, String> {
     }
 
     public void delete(String id) {
-        authorRepository.findById(id).ifPresent(author -> {
-            authorRepository.delete(author);
-            log.debug("Deleted Author: {}", author);
-        });
+        Query query = new Query();
+        query.addCriteria(where("authorList.id").is(id));
+        if (mongoTemplate.count(query, Book.class) > 0) {
+            throw new CustomizeException(Constants.ERR_DELETE);
+        } else {
+            authorRepository.findById(id).ifPresent(author -> {
+                authorRepository.delete(author);
+                log.debug("Deleted Author: {}", author);
+            });
+        }
+    }
+
+    public List<Author> findByFullNameIsLike(String fullName) {
+        return authorRepository.findByFullNameIsLike(fullName);
     }
 }
