+++ books/src/main/java/soft/co/books/domain/service/TopicService.java
@@ -6,9 +6,15 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.mongodb.core.MongoTemplate;
+import org.springframework.data.mongodb.core.query.Criteria;
 import org.springframework.data.mongodb.core.query.Query;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import soft.co.books.configuration.Constants;
 import soft.co.books.configuration.database.CustomBaseService;
+import soft.co.books.configuration.error.CustomizeException;
+import soft.co.books.domain.collection.Book;
+import soft.co.books.domain.collection.Magazine;
 import soft.co.books.domain.collection.Topic;
 import soft.co.books.domain.repository.TopicRepository;
 import soft.co.books.domain.service.dto.PageResultDTO;
@@ -23,6 +29,7 @@ import static org.springframework.data.mongodb.core.query.Criteria.where;
  * Service class for managing topics.
  */
 @Service
+@Transactional
 public class TopicService extends CustomBaseService<Topic, String> {
 
     private final Logger log = LoggerFactory.getLogger(TopicService.class);
@@ -45,8 +52,12 @@ public class TopicService extends CustomBaseService<Topic, String> {
         query.skip(pageable.getPageNumber() * pageable.getPageSize());
         query.with(pageable.getSort());
 
-        if (topicDTO.getName() != null && !topicDTO.getName().isEmpty())
-            query.addCriteria(where("name").regex(topicDTO.getName()));
+        if (topicDTO.getName() != null && !topicDTO.getName().isEmpty()) {
+            Criteria criteria = new Criteria();
+            criteria.orOperator(where("spanishName").regex(topicDTO.getName(), "i"),
+                    where("englishName").regex(topicDTO.getName(), "i"));
+            query.addCriteria(criteria);
+        }
 
         Page<Topic> topics = new PageImpl<>(mongoTemplate.find(query, Topic.class));
         resultDTO.setElements(topics.stream().map(TopicDTO::new).collect(Collectors.toList()));
@@ -56,7 +67,8 @@ public class TopicService extends CustomBaseService<Topic, String> {
 
     public Optional<TopicDTO> createTopic(TopicDTO topicDTO) {
         Topic topic = new Topic();
-        topic.setName(topicDTO.getName());
+        topic.setEnglishName(topicDTO.getEnglishName());
+        topic.setSpanishName(topicDTO.getSpanishName());
         topic.setId(topicDTO.getId());
 
         log.debug("Created Information for Topic: {}", topic);
@@ -76,7 +88,8 @@ public class TopicService extends CustomBaseService<Topic, String> {
                 .filter(Optional::isPresent)
                 .map(Optional::get)
                 .map(topic -> {
-                    topic.setName(topicDTO.getName());
+                    topic.setEnglishName(topicDTO.getEnglishName());
+                    topic.setSpanishName(topicDTO.getSpanishName());
                     topicRepository.save(topic);
                     log.debug("Changed Information for Topic: {}", topic);
                     return topic;
@@ -85,9 +98,15 @@ public class TopicService extends CustomBaseService<Topic, String> {
     }
 
     public void delete(String id) {
-        topicRepository.findById(id).ifPresent(topic -> {
-            topicRepository.delete(topic);
-            log.debug("Deleted Editor: {}", topic);
-        });
+        Query query = new Query();
+        query.addCriteria(where("topic.id").is(id));
+        if (mongoTemplate.count(query, Book.class) > 0 || mongoTemplate.count(query, Magazine.class) > 0) {
+            throw new CustomizeException(Constants.ERR_DELETE);
+        } else {
+            topicRepository.findById(id).ifPresent(topic -> {
+                topicRepository.delete(topic);
+                log.debug("Deleted Topic: {}", topic);
+            });
+        }
     }
 }
